+++ AlphaEngine/Utils/Utils.h
@@ -3,7 +3,7 @@
 
 int PosToMapX(float x);
 int PosToMapY(float y);
-float MapToPosX(int ind);
-float MapToPosY(int ind);
+float MapToPosX(float ind);
+float MapToPosY(float ind);
 
 float GetSqDistance(float ax, float ay, float bx, float by);
